Move library to separate package
